+++ lib/navigation/model/routes/classes_feedback_path.dart
@@ -1,0 +1,41 @@
+part of route_paths;
+
+class ClassesFeedbackPath extends RoutePath {
+  ClassesFeedbackPath() : super(ClassFeedbackChecklist.routeName);
+
+  @override
+  Widget get page/*=> const ClassFeedbackChecklist();*/ {
+    return Builder(
+      builder: (BuildContext context) {
+        final authProvider = Provider.of<AuthProvider>(context);
+        final classProvider = Provider.of<ClassProvider>(context);
+
+        if (!authProvider.isAuthenticated || authProvider.isAnonymous) {
+          return ErrorPage(
+            errorMessage: S.current.warningAuthenticationNeeded,
+          );
+        }
+
+        return FutureBuilder<List<ClassHeader>>(
+          future: classProvider.fetchClassHeaders(uid: authProvider.uid),
+          builder: (BuildContext context,
+              AsyncSnapshot<List<ClassHeader>> snapshot) {
+            if (snapshot.connectionState == ConnectionState.done) {
+              if (!snapshot.hasData) {
+                return ErrorPage(
+                  errorMessage: S.current.errorClassCannotBeEmpty,
+                );
+              }
+
+              final headers = snapshot.data.toSet();
+              return ClassFeedbackChecklist(
+                classes: headers,
+              );
+            }
+            return const CircularProgressIndicator();
+          },
+        );
+      },
+    );
+  }
+}
