+++ lib/navigation/model/routes/edit_profile_path.dart
@@ -1,0 +1,8 @@
+part of route_paths;
+
+class EditProfilePath extends RoutePath {
+  EditProfilePath() : super(EditProfilePage.routeName);
+
+  @override
+  Widget get page => const EditProfilePage();
+}
