+++ lib/navigation/model/routes/faq_path.dart
@@ -1,0 +1,8 @@
+part of route_paths;
+
+class FaqPath extends RoutePath {
+  FaqPath() : super(FaqPage.routeName);
+
+  @override
+  Widget get page => FaqPage();
+}
