+++ lib/navigation/model/routes/filter_path.dart
@@ -1,0 +1,8 @@
+part of route_paths;
+
+class FilterPath extends RoutePath {
+  FilterPath() : super(FilterPage.routeName);
+
+  @override
+  Widget get page => const FilterPage();
+}
