+++ lib/navigation/model/routes/home_path.dart
@@ -1,0 +1,8 @@
+part of route_paths;
+
+class HomePath extends RoutePath {
+  HomePath() : super(HomePage.routeName);
+
+  @override
+  Widget get page => const HomePage();
+}
