+++ lib/navigation/model/routes/login_path.dart
@@ -1,0 +1,8 @@
+part of route_paths;
+
+class LoginPath extends RoutePath {
+  LoginPath() : super(LoginView.routeName);
+
+  @override
+  Widget get page => LoginView();
+}
