+++ lib/navigation/model/routes/news_feed_path.dart
@@ -1,0 +1,8 @@
+part of route_paths;
+
+class NewsFeedPath extends RoutePath {
+  NewsFeedPath() : super(NewsFeedPage.routeName);
+
+  @override
+  Widget get page => NewsFeedPage();
+}
