+++ lib/navigation/model/routes/people_path.dart
@@ -1,0 +1,8 @@
+part of route_paths;
+
+class PeoplePath extends RoutePath {
+  PeoplePath() : super(PeoplePage.routeName);
+
+  @override
+  Widget get page => const PeoplePage();
+}
