+++ lib/navigation/model/routes/portal_path.dart
@@ -1,0 +1,8 @@
+part of route_paths;
+
+class PortalPath extends RoutePath {
+  PortalPath() : super(PortalPage.routeName);
+
+  @override
+  Widget get page => const PortalPage();
+}
