+++ lib/navigation/model/routes/request_permissions_path.dart
@@ -1,0 +1,8 @@
+part of route_paths;
+
+class RequestPermissionsPath extends RoutePath {
+  RequestPermissionsPath() : super(RequestPermissionsPage.routeName);
+
+  @override
+  Widget get page => RequestPermissionsPage();
+}
