+++ lib/navigation/model/routes/root_path.dart
@@ -1,0 +1,8 @@
+part of route_paths;
+
+class RootPath extends RoutePath {
+  RootPath() : super('/');
+
+  @override
+  Widget get page => AppLoadingScreen();
+}
