+++ lib/navigation/model/routes/settings_path.dart
@@ -1,0 +1,8 @@
+part of route_paths;
+
+class SettingsPath extends RoutePath {
+  SettingsPath() : super(SettingsPage.routeName);
+
+  @override
+  Widget get page => SettingsPage();
+}
