+++ lib/navigation/model/routes/signup_path.dart
@@ -1,0 +1,8 @@
+part of route_paths;
+
+class SignUpPath extends RoutePath {
+  SignUpPath() : super(SignUpView.routeName);
+
+  @override
+  Widget get page => SignUpView();
+}
