+++ lib/navigation/model/routes/timetable_path.dart
@@ -1,0 +1,8 @@
+part of route_paths;
+
+class TimetablePath extends RoutePath {
+  TimetablePath() : super(TimetablePage.routeName);
+
+  @override
+  Widget get page => const TimetablePage();
+}
